Fix size of atoms, size_t parameters. Addresses CA1901

+++ Rubberduck.Core/Common/Hotkeys/HotkeyInfo.cs
@@ -9,13 +9,13 @@ namespace Rubberduck.Common.Hotkeys
     {
         private const Keys Modifiers = Keys.Alt | Keys.Control | Keys.Shift;
 
-        public HotkeyInfo(IntPtr hookId, Keys keys)
+        public HotkeyInfo(ushort hookId, Keys keys)
         {
             HookId = hookId;
             Keys = keys;
         }
 
-        public IntPtr HookId { get; }
+        public ushort HookId { get; }
         public Keys Keys { get; }
 
         public override string ToString()
